refactor: extract message builders into shared helper
